Revert "Split openstack-glance into new subpackage"
This reverts commit 8435b033f14cdabc7a345f81abe314acf49aaf81.

This patch broke upstream puppet, and TripleO badly and
will likely break other end users as well.

If we do this we absolutely need to consider the effect on
end users who may be using existing Glance packages.

Glance (upstream) hasn't changed for quite some time
so this split now seems kind of arbitrary to me.

Change-Id: Iae0afb982973f89634b422debd900e658420a54e