Update patch set 1
Patch Set 1:

> So, I'm not saying this patch is not good; Jenkins CI is happy, so
 > it probably fixes our issues. But I do not understand:
 > 
 > 1) How the upstream patch is related to the failures we had in RDO
 > 2) How adding a comment in the spec file solves the aforementioned
 > issue.
 > 
 > Could you enlighten me? To be fair I'm not super familiar with RDO,
 > and I'm certainly no expert at RPM packaging.

All these reviews created by the "rdo-trunk" user are automatically generated when any of the DLRN workers finds that a package failed to build. The added comment in the spec is just a way to have some change that allows the review to be opened, and it needs to be removed when the actual fix is added. [1] has a good explanation of the workflow.

In this case, we have a single distgit branch (rpm-master) shared by more than one worker. That creates an issue for the CI job, because it has to decide which configuration to use for the CI checks based on the ZUUL_BRANCH parameter alone. Long story short, the CI +1 is a red herring in this case, and doesn't mean the spec change fixes anything. It just means that we need to investigate what is going on.

[1] https://blogs.rdoproject.org/7903/how-to-fix-a-ftbfs-using-dlrn

Patch-set: 1