Update patch set 1
Patch Set 1:

OK, thanks!

"The added comment in the spec is just a way to have some change that allows the review to be opened, and it needs to be removed when the actual fix is added."

OK, so the actual fix will be proposed as patchset 2 of this review, right?

Patch-set: 1
Reviewer: Gerrit User 255 <255@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0