Remove privsep-helper from sudo config
This was for os-brick's use of oslo.privsep
and is only used when using the Cinder store
for Glance.

This functionality is not working yet in RDO
Ocata packages due to other changes that are
needed, and use of "privsep-helper *" was
flagged as a potentially dangerous command
to allow here.  We will work on narrowing this
configuration down to be more specific for Pike.

Change-Id: Ia33ce485b1fd50d9a568f71821b152c66de7ebf3
(cherry picked from commit 3db241c65c7f43acb465c70b13c9e412a3f0bd0a)