Update patch set 1
Patch Set 1:

I have run some local tests. The difference between upstream and our environment is that upstream uses sphinx 1.8.2 and we use 1.6.5.

With 1.8.2, this warning does not cause an error, but with 1.6.5 it does. I propose to simply remove -W from the sphinx-build command line, I've tested it and it works.

Patch-set: 1
Label: Code-Review=0