Define the default oslo.coordination lock_path
The recent os-brick requires the lock_path parameter when cinder
backend is used[1], to avoid conflicting management of volume
connections. Because of this change now glance can't start if
the parameter is not set but cinder backend is used.

This change introduces the default value as is done for the other
services like cinder to keep glance "functional" even with that
new behavior in os-brick.

Users still need to set a reasonable and common default value according
to their deployment setup, but that is also applicable to any other
services collocated with glance, for example cinder, so avoid direct
breakage is more preferable option.

Change-Id: I2c7d076a01300be51d08c5a9bb8fbe4078164756